extract responsibilities from products
