+++ product-service/src/libs/products/index.ts
@@ -1,0 +1,17 @@
+import {
+  getProducts,
+  getProductById,
+  createProduct,
+} from './products-service';
+
+import {
+  Product,
+} from './products';
+
+
+export {
+  Product,
+  getProducts,
+  getProductById,
+  createProduct,
+};
